Put the text stuff into separate files
